==> 0_setup.R <==
## 0 - setup

wd <- "XC_vigna"
# install required R packages 

install.packages(c("terra", "geodata", "igraph", "remotes"))

remotes::install_github("https://github.com/MEEreview/MEEreview_exsitu")
remotes::install_github("https://github.com/MEEreview/MEEreview_data")
